Hi @KimPiks, I've tried to make some performance optimizations to this repo. Following are the benchmark results before and after the changes:

**Benchmarks before changes:**
|                  Method |         Mean |        Error |       StdDev |        Median |
|------------------------ |-------------:|-------------:|-------------:|--------------:|
|          IsEmailAddress |  4,500.00 ns |   260.813 ns |   769.013 ns |  4,325.969 ns |
| IsEmailAddressException | 15,339.99 ns | 1,017.827 ns | 3,001.084 ns | 14,791.098 ns |
|           IsPhoneNumber |    356.76 ns |    24.009 ns |    70.790 ns |    332.227 ns |
|  IsPhoneNumberException | 11,005.09 ns |   586.765 ns | 1,730.091 ns | 10,522.444 ns |
|               MinLength |     10.20 ns |     0.348 ns |     1.025 ns |      9.943 ns |
|      MinLengthException | 11,749.40 ns |   766.834 ns | 2,261.028 ns | 10,663.509 ns |
|               MaxLength |     13.55 ns |     0.937 ns |     2.764 ns |     12.915 ns |
|      MaxLengthException | 11,573.79 ns |   564.759 ns | 1,665.206 ns | 11,027.770 ns |
|                 IsEmpty |     12.63 ns |     0.587 ns |     1.729 ns |     12.104 ns |
|        IsEmptyException | 12,306.88 ns |   824.949 ns | 2,432.380 ns | 11,441.824 ns |
|              IsNotEmpty |     11.67 ns |     0.141 ns |     0.131 ns |     11.674 ns |
|     IsNotEmptyException | 10,710.99 ns |   758.397 ns | 2,236.149 ns |  9,951.210 ns |
|                    Only |    237.58 ns |    10.973 ns |    32.354 ns |    226.122 ns |
|           OnlyException | 10,409.26 ns |   207.628 ns |   568.379 ns | 10,327.890 ns |
|                Contains |    326.60 ns |     6.556 ns |    14.528 ns |    324.911 ns |
|       ContainsException | 10,890.89 ns |   437.388 ns | 1,289.648 ns | 10,447.099 ns |

**Benchmarks after changes:**
|                  Method |          Mean |       Error |        StdDev |        Median |
|------------------------ |--------------:|------------:|--------------:|--------------:|
|          IsEmailAddress |  1,188.387 ns |  51.1407 ns |   150.7894 ns |  1,151.605 ns |
| IsEmailAddressException | 12,769.349 ns | 920.8283 ns | 2,715.0832 ns | 12,073.502 ns |
|           IsPhoneNumber |    140.766 ns |   7.2142 ns |    21.2713 ns |    135.719 ns |
|  IsPhoneNumberException | 11,167.433 ns | 659.0172 ns | 1,943.1270 ns | 10,281.595 ns |
|               MinLength |     10.177 ns |   0.2652 ns |     0.7820 ns |      9.934 ns |
|      MinLengthException | 10,372.014 ns | 409.2473 ns | 1,206.6750 ns | 10,011.041 ns |
|               MaxLength |      9.676 ns |   0.1643 ns |     0.1537 ns |      9.601 ns |
|      MaxLengthException |  9,645.725 ns | 192.7296 ns |   380.4294 ns |  9,554.854 ns |
|                 IsEmpty |     11.464 ns |   0.2648 ns |     0.5644 ns |     11.337 ns |
|        IsEmptyException |  9,153.734 ns | 330.2016 ns |   973.6070 ns |  8,810.167 ns |
|              IsNotEmpty |     11.981 ns |   0.2785 ns |     0.4729 ns |     11.910 ns |
|     IsNotEmptyException |  9,103.997 ns | 180.6945 ns |   497.6853 ns |  8,966.581 ns |
|                    Only |    231.369 ns |   7.0798 ns |    20.8749 ns |    225.519 ns |
|           OnlyException | 10,263.493 ns | 241.5819 ns |   712.3097 ns | 10,035.780 ns |
|                Contains |    315.046 ns |   6.3175 ns |    16.3074 ns |    310.848 ns |
|       ContainsException | 10,490.389 ns | 255.8929 ns |   754.5060 ns | 10,284.167 ns |

As you can see the Email and PhoneNumber benchmark durations have gone down. I've also made sure that all the unit tests pass following these changes as well. Could you please confirm whether or not you think my changes are valid?

Thank you!